+++ CHANGELOG.md
@@ -1,4 +1,14 @@
-# [2.2.0]
+# [2.3.0]
+
+- Navigation has accept Dynamic type for will be return this result
+- Removed special character extension added on string.
+
+## [2.2.1]
+
+- Added custom launch options
+- String extension fixed for nullable.
+
+## [2.2.0]
 
 - Lint updated then fixed problems.
 - @desxz added file type, tablet and new extensions to library.